Read in temperature from Gromacs files.
Add a function to read in temperature from Gromacs dhdl.xvg files, so the user
does not need to specify it directly. The temperature is read in from one of the
files, whichever is first in the pre-sorted list (as they should be the same
anyway). The temperature read in overrides the default temp and anything
specified on the command line. If there is a problem reading in the temperature
from dhdl.xvg (I can't see how at this point) it will fall back to the default
or the one specified on the command line.